Delete client directory
I already have a front end, this front end was created with create-react-app and has better UI and updates, I am hoping to fix it and then replace my current front end with the npm build file from this set. idk if there's a better way to fork and play around, so for now we're running analog.